Exercise on the push button phone
An image of the first design of the push button phone.
The students should list the differences with a modern phone.
The most important difference is between the 5x2 keypad of the original
and the 3x4 keypad of a modern phone. The students should consider how
conventions in user interfaces are needed and used.